[CL-1252] Consolidate to use one messages file
