--- front/app/modules/free/native_surveys/admin/components/SurveyBuilderSettings/messages.ts
@@ -1,28 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  close: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.settings.close',
-    defaultMessage: 'Close',
-  },
-  delete: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.settings.delete',
-    defaultMessage: 'Delete',
-  },
-  shortAnswer: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.settings.shortAnswer',
-    defaultMessage: 'Short answer',
-  },
-  questionTitle: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.settings.questionTitle',
-    defaultMessage: 'Question title',
-  },
-  questionDescription: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.settings.questionDescription',
-    defaultMessage: 'Question description',
-  },
-  required: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.settings.required',
-    defaultMessage: 'Required',
-  },
-});
