--- front/app/modules/free/native_surveys/admin/components/SurveyBuilderToolbox/messages.ts
@@ -1,16 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  addSurveyContent: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.toolbox.addSurveyContent',
-    defaultMessage: 'Add survey content',
-  },
-  shortAnswer: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.toolbox.shortAnswer',
-    defaultMessage: 'Short answer',
-  },
-  newField: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.toolbox.newField',
-    defaultMessage: 'New field',
-  },
-});
