--- front/app/modules/free/native_surveys/admin/components/SurveyBuilderTopBar/messages.ts
@@ -1,16 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  survey: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.topBar.surveyText',
-    defaultMessage: 'Survey',
-  },
-  viewSurvey: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.topBar.viewSurvey',
-    defaultMessage: 'View survey',
-  },
-  save: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.topBar.save',
-    defaultMessage: 'Save',
-  },
-});
