+++ front/app/modules/free/native_surveys/admin/components/Tab.tsx
@@ -17,7 +17,7 @@ const Tab: FC<Props & InjectedIntlProps> = ({
   useEffect(() => {
     onData({
       configuration: {
-        label: formatMessage(messages.nativeSurveyTab),
+        label: formatMessage(messages.surveyText),
         url: 'native-survey',
         feature: 'native_surveys',
         name: 'nativeSurveys',
