+++ front/app/modules/free/native_surveys/admin/components/messages.ts
@@ -1,8 +1,56 @@
 import { defineMessages } from 'react-intl';
 
 export default defineMessages({
-  nativeSurveyTab: {
-    id: 'app.containers.AdminPage.ProjectEdit.nativeSurveyTab',
+  surveyText: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.surveyText',
     defaultMessage: 'Survey',
+  },
+  viewSurvey: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.viewSurvey',
+    defaultMessage: 'View survey',
+  },
+  save: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.save',
+    defaultMessage: 'Save',
+  },
+  addSurveyContent: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.addSurveyContent',
+    defaultMessage: 'Add survey content',
+  },
+  shortAnswer: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.shortAnswer',
+    defaultMessage: 'Short answer',
+  },
+  newField: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.newField',
+    defaultMessage: 'New field',
+  },
+  close: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.close',
+    defaultMessage: 'Close',
+  },
+  delete: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.delete',
+    defaultMessage: 'Delete',
+  },
+  questionTitle: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.questionTitle',
+    defaultMessage: 'Question title',
+  },
+  questionDescription: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.questionDescription',
+    defaultMessage: 'Question description',
+  },
+  required: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.required',
+    defaultMessage: 'Required',
+  },
+  deleteButtonLabel: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.deleteButtonLabel',
+    defaultMessage: 'Delete',
+  },
+  editButtonLabel: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.editButtonLabel',
+    defaultMessage: 'Edit',
   },
 });
