--- front/app/modules/free/native_surveys/components/SurveyFields/messages.ts
@@ -1,20 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  required: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.customFields.required',
-    defaultMessage: 'Required',
-  },
-  deleteButtonLabel: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.customFields.deleteButtonLabel',
-    defaultMessage: 'Delete',
-  },
-  editButtonLabel: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.customFields.editButtonLabel',
-    defaultMessage: 'Edit',
-  },
-  shortAnswer: {
-    id: 'app.containers.AdminPage.ProjectEdit.survey.customFields.shortAnswer',
-    defaultMessage: 'Short answer',
-  },
-});
